HW6/requirements.txt
google-generativeai>=0.8.0
pdf2image>=1.17.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
edge-tts>=7.0.0
